Implemented: Anti-Aliasing.
git-svn-id: http://foreverbell.0ginr.com/svn/ray_tracer@32 a142d196-329d-4e72-972a-086f8ffe8ca1

+++ ray_tracer/camera.hpp
@@ -10,7 +10,7 @@ namespace ray_tracer {
 	class camera {
 	public:
 		virtual ~camera() = 0;
-		virtual colorRGB render_scene(int, int, int, int, world *) const = 0;
+		virtual colorRGB render_scene(double, double, int, int, world *) const = 0;
 		point3D get_eye() const;
 		void set_eye(const point3D &);
 		point3D get_lookat() const;
