+++ ray_tracer/camera_orthographic.hpp
@@ -11,7 +11,7 @@ namespace ray_tracer {
 	public:
 		camera_orthographic();
 		camera_orthographic(const point3D &, const point3D &, const vector3D &);
-		colorRGB render_scene(int, int, int, int, world *) const;
+		colorRGB render_scene(double, double, int, int, world *) const;
 	};
 }
